+++ bin/evinse.js
@@ -8,14 +8,18 @@ import { load as _load } from "js-yaml";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 import {
+  analyzeProject,
+  createEvinseFile,
+  prepareDB,
+} from "../lib/enviser/index.js";
+import {
   printCallStack,
   printOccurrences,
   printReachables,
   printServices,
-} from "../display.js";
-import { analyzeProject, createEvinseFile, prepareDB } from "../evinser.js";
-import { ATOM_DB } from "../utils.js";
-import { validateBom } from "../validator.js";
+} from "../lib/helpers/display.js";
+import { ATOM_DB } from "../lib/helpers/utils.js";
+import { validateBom } from "../lib/helpers/validator.js";
 
 // Support for config files
 const configPath = findUpSync([
